moved manipulator to facade
